+++ tools/opencv-test/src/encode/av1.rs
@@ -1,0 +1,32 @@
+use super::Args;
+use anyhow::Result;
+
+use std::{
+    fs::OpenOptions,
+    io::{BufWriter, Write},
+};
+
+use opencv::{prelude::*, videoio};
+
+pub fn encode_av1(args: Args) -> Result<()> {
+    let cam = videoio::VideoCapture::from_file(&args.input, videoio::CAP_ANY)?;
+    let opened = videoio::VideoCapture::is_opened(&cam)?;
+    if !opened {
+        panic!("Unable to open video file!");
+    }
+
+    // https://docs.opencv.org/3.4/d4/d15/group__videoio__flags__base.html
+    let frame_width = cam.get(3)? as u32;
+    let frame_height = cam.get(4)? as u32;
+    let fps = cam.get(5)? as f32;
+
+    let output_file = OpenOptions::new()
+        .read(false)
+        .write(true)
+        .create(true)
+        .truncate(true)
+        .open(args.output)?;
+    let mut writer = BufWriter::new(output_file);
+
+    todo!()
+}
