+++ tools/opencv-test/src/utils.rs
@@ -139,3 +139,66 @@ impl YUVSource for YUVBuf {
         self.width as i32 / 2
     }
 }
+
+// attempts to avoid the loss when converting from BGR to YUV420 by quadrupling the size of the output. this ensures no UV samples are discarded/averaged
+// unfortunately is still lossy
+pub fn bgr_to_yuv_lossy(s: &[u8], width: usize, height: usize) -> Vec<u8> {
+    // for y
+    let y_rc_2_idx = |row: usize, col: usize| (row * width * 2) + col;
+
+    let get_y = |rgb: (f32, f32, f32)| {
+        // best. appears to be from the wikipedia page on YCbCr
+        (0.2578125 * rgb.0 + 0.50390625 * rgb.1 + 0.09765625 * rgb.2 + 16.0) as u8
+        // full scale:   // https://web.archive.org/web/20180423091842/http://www.equasys.de/colorconversion.html
+        //(0.299 * rgb.0 + 0.587 * rgb.1 + 0.114 * rgb.2 + 0.0) as u8
+        // hdtv
+        //(0.183 * rgb.0 + 0.614 * rgb.1 + 0.062 * rgb.2 + 16.0) as u8
+    };
+
+    let get_u = |rgb: (f32, f32, f32)| {
+        (-0.1484375 * rgb.0 + -0.2890625 * rgb.1 + 0.4375 * rgb.2 + 128.0) as u8
+        //(-0.169 * rgb.0 + -0.331 * rgb.1 + 0.500 * rgb.2 + 128.0) as u8
+        //(-0.101 * rgb.0 + -0.339 * rgb.1 + 0.439 * rgb.2 + 128.0) as u8
+    };
+
+    let get_v = |rgb: (f32, f32, f32)| {
+        (0.4375 * rgb.0 + -0.3671875 * rgb.1 + -0.0703125 * rgb.2 + 128.0) as u8
+        //(0.500 * rgb.0 + -0.419 * rgb.1 + -0.081 * rgb.2 + 128.0) as u8
+        //(0.439 * rgb.0 + -0.399 * rgb.1 + -0.040 * rgb.2 + 128.0) as u8
+    };
+
+    let yuv_len = (width * height) * 6;
+    let mut yuv: Vec<u8> = Vec::new();
+    yuv.resize(yuv_len, 0);
+    let u_base = (width * height) * 4;
+    let v_base = u_base + (width * height);
+    let mut uv_idx = 0;
+    for row in 0..height {
+        for col in 0..width {
+            let base_pos = (col + row * width) * 3;
+            let b = s[base_pos];
+            let g = s[base_pos + 1];
+            let r = s[base_pos + 2];
+
+            let rgb = (r as _, g as _, b as _);
+            let (y, u, v) = (get_y(rgb), get_u(rgb), get_v(rgb));
+
+            // each byte in the u/v plane corresponds to a 4x4 square on the y plane
+            let y_row = row * 2;
+            let y_col = col * 2;
+
+            let idx = y_rc_2_idx(y_row, y_col);
+            yuv[idx] = y;
+            yuv[idx + 1] = y;
+            let idx = y_rc_2_idx(y_row + 1, y_col);
+            yuv[idx] = y;
+            yuv[idx + 1] = y;
+
+            yuv[u_base + uv_idx] = u;
+            yuv[v_base + uv_idx] = v;
+            uv_idx += 1;
+        }
+    }
+
+    yuv
+}
